requirements.txt
asgiref==3.2.10
Django==3.1.6
djangorestframework==3.12.2
Pillow==7.2.0
python-dotenv==0.15.0
pytz==2020.1
sqlparse==0.3.1
